+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/bulkscan/transformation/BulkScanFormTransformerFactory.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.service.bulkscan.transformations;
+package uk.gov.hmcts.reform.finrem.caseorchestration.service.bulkscan.transformation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
